Refactor protocol into internal, external modules.
This commit just moves things around and patches import paths.
