+++ zebra-network/src/constants.rs
@@ -3,7 +3,7 @@
 use std::time::Duration;
 
 // XXX should these constants be split into protocol also?
-use crate::protocol::types::*;
+use crate::protocol::external::types::*;
 
 /// The timeout for requests made to a remote peer.
 pub const REQUEST_TIMEOUT: Duration = Duration::from_secs(10);
