+++ zebra-network/src/peer/error.rs
@@ -68,7 +68,7 @@ impl ErrorSlot {
 pub enum HandshakeError {
     /// The remote peer sent an unexpected message during the handshake.
     #[error("The remote peer sent an unexpected message: {0:?}")]
-    UnexpectedMessage(crate::protocol::message::Message),
+    UnexpectedMessage(crate::protocol::external::Message),
     /// The peer connector detected handshake nonce reuse, possibly indicating self-connection.
     #[error("Detected nonce reuse, possible self-connection")]
     NonceReuse,
