+++ zebra-network/src/protocol/external.rs
@@ -1,0 +1,12 @@
+/// A Tokio codec that transforms an `AsyncRead` into a `Stream` of `Message`s.
+mod codec;
+/// Inventory items.
+mod inv;
+/// An enum of all supported Bitcoin message types.
+mod message;
+/// Newtype wrappers for primitive types.
+pub mod types;
+
+pub use codec::Codec;
+pub use inv::InventoryHash;
+pub use message::Message;
